+++ driver/aws/s3set/doc.go
@@ -1,0 +1,2 @@
+// Package s3set provides a set store backed by AWS S3.
+package s3set