+++ driver/aws/s3set/store.go
@@ -1,4 +1,4 @@
-package s3kv
+package s3set
 
 import (
 	"context"
@@ -7,10 +7,10 @@ import (
 	"github.com/aws/aws-sdk-go-v2/service/s3"
 	"github.com/dogmatiq/enginekit/x/xsync"
 	"github.com/dogmatiq/persistencekit/driver/aws/internal/s3x"
-	"github.com/dogmatiq/persistencekit/kv"
+	"github.com/dogmatiq/persistencekit/set"
 )
 
-// store is an implementation of [kv.BinaryStore] that persists to an S3 bucket.
+// store is an implementation of [set.BinaryStore] that persists to an S3 bucket.
 type store struct {
 	Client    *s3.Client
 	Bucket    string
@@ -19,13 +19,13 @@ type store struct {
 	createBucketOnce xsync.SucceedOnce
 }
 
-// NewBinaryStore returns a new [kv.BinaryStore] that uses the given S3 client
-// to store key/value pairs in the given bucket.
+// NewBinaryStore returns a new [set.BinaryStore] that uses the given S3 client
+// to store set members in the given bucket.
 func NewBinaryStore(
 	client *s3.Client,
 	bucket string,
 	options ...Option,
-) kv.BinaryStore {
+) set.BinaryStore {
 	if bucket == "" {
 		panic("bucket name must not be empty")
 	}
@@ -48,7 +48,7 @@ type Option func(*store)
 // WithRequestHook is an [Option] that configures fn as a pre-request hook.
 //
 // Before each S3 API request, fn is passed a pointer to the input struct, e.g.
-// [s3.GetObjectInput], which it may modify in-place. It may be called with any
+// [s3.HeadObjectInput], which it may modify in-place. It may be called with any
 // S3 request type. The types of requests used may change in any version without
 // notice.
 //
@@ -60,21 +60,19 @@ func WithRequestHook(fn func(any) []func(*s3.Options)) Option {
 	}
 }
 
-// Open returns the keyspace with the given name.
-func (s *store) Open(ctx context.Context, name string) (kv.BinaryKeyspace, error) {
+// Open returns the set with the given name.
+func (s *store) Open(ctx context.Context, name string) (set.BinarySet, error) {
 	if err := s.createBucketOnce.Do(ctx, s.createBucket); err != nil {
 		return nil, err
 	}
 
-	ks := &keyspace{
+	return &setimpl{
 		client:          s.Client,
 		onRequest:       s.OnRequest,
 		name:            name,
 		bucket:          s.Bucket,
-		objectKeyPrefix: "kv/" + url.PathEscape(name) + "/",
-	}
-
-	return ks, nil
+		objectKeyPrefix: "set/" + url.PathEscape(name) + "/",
+	}, nil
 }
 
 func (s *store) createBucket(ctx context.Context) error {